+++ src/pokeapi/items.rs
@@ -11,10 +11,10 @@ use crate::{impl_id_and_named, set_endpoint};
 #[cfg_attr(debug_assertions, serde(deny_unknown_fields))]
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq, Hash)]
 pub struct Item {
-    pub id: i64,
+    pub id: i16,
     pub name: String,
-    pub cost: u64,
-    pub fling_power: Option<u64>,
+    pub cost: u32,
+    pub fling_power: Option<u8>,
     pub fling_effect: Option<NamedAPIResource<ItemFlingEffect>>,
     pub attributes: Vec<NamedAPIResource<ItemAttribute>>,
     pub category: NamedAPIResource<ItemCategory>, // incorrectly documented as ItemCategory
@@ -44,14 +44,14 @@ pub struct ItemHolderPokemon {
 #[cfg_attr(debug_assertions, serde(deny_unknown_fields))]
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq, Hash)]
 pub struct ItemHolderPokemonVersionDetail {
-    pub rarity: u64, // incorrectly documented as string
+    pub rarity: u8, // incorrectly documented as string
     pub version: NamedAPIResource<Version>,
 }
 
 #[cfg_attr(debug_assertions, serde(deny_unknown_fields))]
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq, Hash)]
 pub struct ItemAttribute {
-    pub id: i64,
+    pub id: i16,
     pub name: String,
     pub items: Vec<NamedAPIResource<Item>>,
     pub names: Vec<Name>,
@@ -61,7 +61,7 @@ pub struct ItemAttribute {
 #[cfg_attr(debug_assertions, serde(deny_unknown_fields))]
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq, Hash)]
 pub struct ItemCategory {
-    pub id: i64,
+    pub id: i16,
     pub name: String,
     pub items: Vec<NamedAPIResource<Item>>,
     pub names: Vec<Name>,
@@ -71,7 +71,7 @@ pub struct ItemCategory {
 #[cfg_attr(debug_assertions, serde(deny_unknown_fields))]
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq, Hash)]
 pub struct ItemFlingEffect {
-    pub id: i64,
+    pub id: i16,
     pub name: String,
     pub effect_entries: Vec<Effect>,
     pub items: Vec<NamedAPIResource<Item>>, // incorrectly documented as NamedAPIResource
@@ -80,7 +80,7 @@ pub struct ItemFlingEffect {
 #[cfg_attr(debug_assertions, serde(deny_unknown_fields))]
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq, Hash)]
 pub struct ItemPocket {
-    pub id: i64,
+    pub id: i16,
     pub name: String,
     pub categories: Vec<NamedAPIResource<ItemCategory>>,
     pub names: Vec<Name>,
