+++ src/pokeapi/machines.rs
@@ -11,7 +11,7 @@ use crate::{impl_id, set_endpoint};
 #[cfg_attr(debug_assertions, serde(deny_unknown_fields))]
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq, Hash)]
 pub struct Machine {
-    pub id: i64,
+    pub id: i16,
     pub item: NamedAPIResource<Item>,
     #[serde(rename = "move")]
     pub move_: NamedAPIResource<Move>,
